fix(audio): رسائل iOS الصوتية كانت تُسقط تحليل المحادثة كاملة
iOS يرسل audioWaveform كأرقام عشرية 0..1 والخادم يخزّنها كما هي ([Number])،
بينما كان الأندرويد يعرّفها List<Int> و audioDuration كـ Int. فقيمة مثل 0.74
ترمي JsonDataException من Moshi، والنتيجة ليست فقدان الرسالة الصوتية بل:

- جلب المحادثة: يفشل تحليل الرد كاملاً فتبدو المحادثة فارغة
- عبر السوكِت: parseMessage يبتلع الاستثناء ويُرجع null فتختفي الرسالة بصمت

الحل استقبال الحقلين كـ Double (يقبل الصحيح والعشري معاً).

+++ app/src/main/java/com/chathala/hala/feature/chats/data/MessageApiModels.kt
@@ -51,8 +51,10 @@ data class Message(
     val editedAt: String? = null,
     val createdAt: String? = null,
     // ── audio ──
-    val audioDuration: Int? = null,    // بالثواني
-    val audioWaveform: List<Int>? = null,
+    // ملاحظة: iOS يرسل waveform كقيم عشرية 0..1 والمدّة قد تُخزَّن كرقم عشري في الخادم،
+    // لذا نستقبلهما كـ Double حتى لا يفشل تحليل Moshi ويسقط الرد كاملاً.
+    val audioDuration: Double? = null,    // بالثواني
+    val audioWaveform: List<Double>? = null,
     // ── image (disappearing) ──
     val imageSource: String? = null,   // camera | gallery
     val disappearing: DisappearingInfo? = null,
